Bug fix for runLORASR.Functions.CreateLogFile not creating new file
Previously, if a .old log file was left in the working directory,
the FileMove command would fail (with no logging, as this is the
create log function) and the log would be appended to the previous
log.

Note this wouldn't be a problem in a clean folder after running the
tidying functions, because the .old file would have been removed. It
was only discovered while developing the new results section.