Created runLORASR.Results and collateLORASR
Separated out collation of results into a new library and a new
executable. This allows going back over previous results and
collating the results into a batch file.

This came about because the previous transmission results were looking
in the wrong place. This commit doesn't fix that problem, it just
separates out the functionality to make the fix possible to apply
retroactively to existing results. That's why this files are still
version 0.4.0, not 0.4.1.

Version 0.4.1 will be the fix for the correct transmission calculation.