Load the correct transmission value, also keep the core particles value
Previous code was incorrectly reporting the proportion of particles in
the core as the proportion of particles transmitted.

Version 4.1.1 should fix this, and keeps both values in the batch results
file.

Note that the SaveRunResults function is quite long now, it probably would
be better to have some subfunctions, perhaps like:
 * GetFileContents($sInputFile, $sOutputFile)
 * GetTransmission($sInputData, $sOutputData)
 * GetCoreParticles($sOutputData)
 * GetRunDetails($sOutputFile)